+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Added setUrl method to FakeRequest.
+- Added setMethod method to FakeRequest.
 
 ## [2.0.0] - 2018-05-14
 - **Backward compatibility break**: Moved session handling from FakeApplication to FakeRequest.